python_requirements.txt
absl-py==1.2.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
brotlipy==0.7.0
cachetools==5.2.0
certifi==2021.10.8
ci-sdr==0.0.0
click==8.1.3
ConfigArgParse==1.5.3
ctc-segmentation==1.7.1
cycler==0.11.0
Cython==0.29.30
decorator==5.1.1
Distance==0.1.3
ear==2.1.0
einops==0.4.1
espnet==202205
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
fast-bss-eval==0.1.3
filelock==3.7.1
flatbuffers==1.12
fonttools==4.34.4
g2p-en==2.1.0
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
humanfriendly==10.0
importlib-metadata==4.12.0
inflect==5.6.2
jaconv==0.3
jamo==0.4.1
jiwer==2.3.0
joblib==1.1.0
kaldiio==2.17.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
librosa==0.9.2
llvmlite==0.38.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
multipledispatch==0.6.0
nltk==3.7
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pyparsing==3.0.9
pypinyin==0.44.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-wpe==0.0.1
pytz==2022.1
pyworld==0.3.0
PyYAML==6.0
regex==2022.7.9
requests-oauthlib==1.3.1
resampy==0.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
sentencepiece==0.1.96
SoundFile==0.10.3.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.8.1+cpu
torch-complex==0.4.3
torchaudio==0.8.1
torchvision==0.9.1+cpu
typeguard==2.13.3
typing_extensions==4.3.0
Unidecode==1.3.4
wav2vec2-stt==0.2.0
Wave==0.0.2
wavinfo==1.6.3
Werkzeug==2.2.0
wrapt==1.14.1
zipp==3.8.1
